Move TabHeader and TabPanel components to a common location
