+++ src/sidebar/components/tabs/TabHeader.tsx
