+++ src/sidebar/components/tabs/TabPanel.tsx
